Averages and MQTT
Added function to average reading of the adc accross all the used analogue pins and convert that into pressure. The reading on the arduino and the pressure sensor now line up for the most part. 100 samples are taken. 
Also integrated the code into the mqtt frame work so it will work with the node red server in E424.